novel_rewritier/requirements.txt
openai>=1.6.1
python-dotenv>=1.0.0
elevenlabs>=0.2.26
